Reduce global components usage
In https://github.com/mjmlio/mjml/pull/2060 I want to pass `presets`
without mutating global components. In this diff I changed components
map to be consumed from context in component implementations.
